requirements.txt
flask
python-dotenv
PyMuPDF
openai==0.27.6
gunicorn